+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.6.4 (15 Mar 2024)
 * Fixed: Minor type compatability issue with newer versions of PHP.
-* Improvement: Increment the revision number of CSS and JS snippet when using the 'Reset Caches' debug action. (PRO)
+* Improved: Increment the revision number of CSS and JS snippet when using the 'Reset Caches' debug action. (PRO)
 * Fixed: Undefined array key issue when initiating cloud sync. (PRO)
 * Fixed: Bug preventing downloading a single snippet from a bundle. (PRO)
 * Added: AI generation for all snippet types: HTML, CSS, JS. (PRO)